Tests: replaced test crates; case-insensitive note
Leaving commented out regarding file-insensitivity for now.
Tempted to leave if under the macos cfg flag, but it will be a flakey
test.  As macos *can* set the filesystem to be case sensitive (on disk
format, I believe).  Simultaneously, it's seems notable enough to keep
about for at least now.