Build(fix): forcing cargo-test to run serially
Cargo nextest runs in paralel without issues and cargo test succeeds on
test sindividually without issues.  It's not yet clear what the cause of
the temporary directory collision is.  (But it persists across a
regression to stable.)
Running cargo-test serially is a workaround for now.

Note: serializing with: `cargo test -- --test-threads 1`
NOT with : `cargo test --jobs 1`
`--jobs` is the number of **compilation** jobs
** -- --test-threads** is the number of threads to run the command
accross